+++ src/aux-common/aux-format-2/AuxOpTypes.ts
@@ -1,7 +1,7 @@
 import { assign } from 'lodash';
 import { AtomId, Atom, atom } from '@casual-simulation/causal-trees';
 import { sign, verify, keypair, getHash } from '@casual-simulation/crypto';
-import stringify from 'fast-json-stable-stringify';
+import stringify from '@casual-simulation/fast-json-stable-stringify';
 
 /**
  * The list of operation types.